fixing: handling data null or not found

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -51,7 +51,17 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed find one banner");
+            }
+
             const result = await BannerModel.findById(id);
+
+            if (!result) {
+                return response.notFound(res, "failed find one banner");
+            }
+
             response.success(res, result, "success find one banner");
         } catch(error) {
             response.error(res, error, "failed to find one banner");
@@ -60,6 +70,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed update banner");
+            }
+
             const result = await BannerModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -71,6 +86,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed remove banner");
+            }
+
             const result = await BannerModel.findByIdAndDelete(id, {
                 new: true,
             });
