+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -55,7 +56,17 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed find one category");
+            }
+
             const result = await CategoryModel.findById(id);
+
+            if (!result) {
+                return response.notFound(res, "failed find one category");
+            }
+
             response.success(res, result, "success find one category");
         } catch (error) {
             response.error(res, error, 'failed find one category');
@@ -64,6 +75,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed update category");
+            }
+
             const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -75,6 +91,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed remove category");
+            }
+
             const result = await CategoryModel.findByIdAndDelete(id, {
                 new: true,
             });
