+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import response from "../utils/response";
 import EventModel, { eventDAO, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
     async create(req: IReqUser, res: Response) {
@@ -54,7 +54,17 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed find one event");
+            }
+
             const result = await EventModel.findById(id);
+
+            if (!result) {
+                return response.notFound(res, "failed find one event");
+            }
+
             response.success(res, result, "success find one event");
         } catch (error) {
             response.error(res, error, "failed find one event");
@@ -63,6 +73,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed update event");
+            }
+
             const result = await EventModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -74,6 +89,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed remove event");
+            }
+
             const result = await EventModel.findByIdAndDelete(id, {
                 new: true,
             });
