+++ src/controllers/ticket.controller.ts
@@ -53,7 +53,17 @@ export default {
     async findOne(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed find one ticket");
+            }
+
             const result = await TicketModel.findById(id);
+
+            if (!result) {
+                return response.notFound(res, "failed find one ticket");
+            }
+
             response.success(res, result, "success find one ticket");
         } catch(error) {
             response.error(res, error, "failed to find one ticket")
@@ -62,6 +72,11 @@ export default {
     async update(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed update ticket");
+            }
+
             const result = await TicketModel.findByIdAndUpdate(id, req.body, {
                 new: true,
             });
@@ -73,6 +88,11 @@ export default {
     async remove(req: IReqUser, res: Response) {
         try {
             const { id } = req.params;
+
+            if (!isValidObjectId(id)) {
+                return response.notFound(res, "failed remove ticket");
+            }
+
             const result = await TicketModel.findByIdAndDelete(id, {
                 new: true,
             });
